Improve configurability

+++ lib/rails_commander/app.rb
@@ -14,28 +14,51 @@ module RailsCommander
     end
 
     def start
+      return if running?
+
       Dir.chdir(@path) do
         @pid = Process.spawn(
+          config.env_vars,
           config.cmd_start,
           unsetenv_others: config.unset_env_vars,
-          %i[out err] => [config.log_path, 'w']
+          out: [config.out_path, 'w'],
+          err: [config.err_path, 'w']
         )
+      end
+    end
+
+    def running?
+      return false unless pid
+
+      begin
+        Process.getpgid(pid)
+        true
+      rescue Errno::ESRCH
+        false
       end
     end
 
     def stop
+      return unless running?
+
       Process.kill('TERM', @pid)
     end
 
-    def db_reset
+    def task(task)
       Dir.chdir(@path) do
         pid = Process.spawn(
-          'bundle exec ./bin/rails db:reset',
+          config.env_vars,
+          "bundle exec ./bin/rails #{task}",
           unsetenv_others: config.unset_env_vars,
-          %i[out err] => [config.log_path, 'w']
+          out: [config.out_path, 'w'],
+          err: [config.err_path, 'w']
         )
         Process.wait(pid)
       end
+    end
+
+    def db_reset
+      task('db:reset')
     end
   end
 end
